Auto-tap: Extract part of the castability probe + option to disable for performance
