+++ forge-gui/res/languages/en-US.properties
@@ -110,6 +110,8 @@ cbScaleLarger=Scale Image Larger
 cbRenderBlackCardBorders=Render Black Card Borders
 cbShowActionableHighlights=Highlight Actionable Cards
 cbShowAutoTapPreview=Highlight Auto-tap Cards
+cbManaCastabilityProbe=Castability-aware Auto-pay
+nlManaCastabilityProbe=When paying mana, consider which hand spells stay castable. Disable for faster Auto-pay on slow devices.
 cbShowLinkedExileCards=Show Linked Exile Cards
 lblActionableHighlightColor=Highlight Color (hex)
 cbLargeCardViewers=Use Large Card Viewers
