+++ forge-gui/src/main/java/forge/localinstance/properties/ForgePreferences.java
@@ -203,6 +203,8 @@ public class ForgePreferences extends PreferencesStore<ForgePreferences.FPref> {
 
         ENFORCE_DECK_LEGALITY ("true"),
         PERFORMANCE_MODE ("false"),
+        /** When false, mana Auto-pay skips hand/command castability dry-runs (faster on low-end devices). */
+        MANA_PAYMENT_CASTABILITY_PROBE ("true"),
         FILTERED_HANDS ("false"),
         MULLIGAN_RULE(MulliganDefs.getDefaultRule().name()),
